requirements.txt
black
eth-brownie
flake8
autoflake
isort
